Define track interpolation wrapping

+++ src/main/PAL/main/track/interpolate_track_angle.c
@@ -7,7 +7,8 @@ s32 InterpolateTrackAngle(s32 pointIndex, s32 weight) {
     if (g_TrackPoints == 0 || g_TrackPointCount <= 0) {
         return 0;
     }
-    next = WrapTrackPointIndex(pointIndex + 1);
+    next = WrapTrackPointIndex((s32)((u32)pointIndex + 1U));
 
-    return BlendAngle(TrackPoint(pointIndex)->angle, TrackPoint(next)->angle, weight);
+    return BlendAngle(TrackPoint(pointIndex)->angle, TrackPoint(next)->angle,
+                      weight);
 }
